Finish tests for `Workout`
